Tests/LibWeb/Text/expected/css/transform-box-change-updates-painted-position.txt

left with transform-box: view-box: 208
left with transform-box: content-box: 108
